% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcoclust.R
\name{rcoclust}
\alias{rcoclust}
\title{Randomized spectral co-clustering using random sampling or random projection}
\usage{
rcoclust(
  A,
  method = c("rsample", "rproject"),
  ky,
  kz,
  p = 10,
  q = 2,
  dist = "normal",
  P,
  normalize = FALSE,
  iter.max = 50,
  nstartkmedian = 10,
  nstartkmeans = 10,
  ...
)
}
\arguments{
\item{A}{The adjacency matrix of a directed network with type "dgCMatrix".}

\item{method}{The method for computing the randomized SVD. Random sampling-based SVD
is implemented if \code{method="rsample"}, and random projection-based
SVD is implemented if \code{method="rproject"}.}

\item{ky}{The number of target row clusters.}

\item{kz}{The number of target column clusters.}

\item{p}{The oversampling parameter in the random projection scheme. Requested only
if \code{method="rproject"}. Default is 10.}

\item{q}{The power parameter in the random projection scheme. Requested only if
\code{method="rproject"}. Default is 2.}

\item{dist}{The distribution of the entry of the random test matrix in the random
projection scheme. Requested only if \code{method="rproject"}. Default
is \code{"normal"}.}

\item{P}{The sampling probability in the random sampling scheme. Requested only
if \code{method="rsample"}.}

\item{normalize}{The k-means is implemented if \code{FALSE}. Otherwise the spherical
k-median is implemented. Default is \code{FALSE}.}

\item{iter.max}{Maximum number of iterations in the \code{\link[stats]{kmeans}} and
that for choosing the starting point of \code{\link[Gmedian]{kGmedian}}.
Default is 50.}

\item{nstartkmedian}{The number of times the k-median algorithm is ran in the
\code{\link[Gmedian]{kGmedian}}. Default is 10.}

\item{nstartkmeans}{The number of random sets in \code{\link[stats]{kmeans}} and that
for choosing the starting point of \code{\link[Gmedian]{kGmedian}}. Default is 10.}

\item{...}{Additional arguments.}
}
\value{
\item{cluster.y}{The row cluster vector (from \code{1:ky}) with the numbers indicating which
             row cluster each node is assigned.}
        \item{cluster.z}{The column cluster vector (from \code{1:kz}) with the numbers indicating which
             column cluster each node is assigned.}
        \item{rvectors.y}{The randomized left \code{min(ky,kz)} singular vectors computed by
             \code{\link[RandClust]{rsvd.pro}} or \code{\link[RandClust]{rsvd.sam}}.}
        \item{rvectors.z}{The randomized right \code{min(ky,kz)} singular vectors computed by
             \code{\link[RandClust]{rsvd.pro}} or \code{\link[RandClust]{rsvd.sam}}.}
}
\description{
Randomized spectral co-clustering for directed networks. The row clusters and
column clusters are computed using two random schemes, namely, the random sampling
and the random projection scheme. Can deal with very large networks.
}
\details{
This function computes the row clusters and column clusters of directed networks using
randomized spectral co-clustering algorithms. The random projection-based SVD or the
random sampling-based SVD is first computed for the adjacency matrix of the directed network.
The k-means or the spherical k-median is then performed on the randomized singular vectors.
}
\examples{
n <- 120
ky <- 2
kz <- 3
cluster.y <- rep(1:ky, each = n/ky)
cluster.z <- rep(1:kz, each = n/kz)
probmat <- matrix(0.2, ky, kz)
diag(probmat) <- 0.1
A <- sample_scbm(type = "scbm", cluster.y, cluster.z, probmat = probmat, graph = FALSE)
rcoclust(A, method ="rsample", ky, kz, P = 0.7, normalize = FALSE)
rcoclust(A, method ="rproject", ky, kz, normalize = TRUE)


}
\seealso{
\code{\link[RandClust]{rsvd.pro}}, \code{\link[RandClust]{rsvd.sam}}.
}
