% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rsample}
\alias{rsample}
\title{Sample a sparse matrix}
\usage{
rsample(A, P)
}
\arguments{
\item{A}{A sparse matrix of type "dgCMatrix".}

\item{P}{The probability that each edge is kept.}
}
\value{
A binary sparse matrix of type "dgCMatrix".
}
\description{
Sample a sparse matrix
}
\examples{
library(Matrix)
set.seed(123)
n = 20
A = matrix(rbinom(n^2, 1, 0.5), 20, 20)
diag(A) = 0
A = as(A, "dgCMatrix")
A
rsample(A, 0.5)
}
