% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rsample_sym}
\alias{rsample_sym}
\title{Sample a symmetric sparse matrix}
\usage{
rsample_sym(A, P, use_lower = TRUE)
}
\arguments{
\item{A}{A sparse matrix of type "dgCMatrix". \code{A} does not need
to be symmetric, see the parameter \code{use_lower}.}

\item{P}{The probability that each edge is kept.}

\item{use_lower}{If \code{TRUE}/\code{FALSE}, only the lower/upper triangular
part of \code{A} is used for sampling.}
}
\value{
A lower triangular, binary, and sparse matrix of type "dgCMatrix".
        The diagonal elements are all zeros.
}
\description{
Sample a symmetric sparse matrix
}
\examples{
library(Matrix)
set.seed(123)
n = 20
A = matrix(rbinom(n^2, 1, 0.5), 20, 20)
A = as(A, "dgCMatrix")
A
rsample_sym(A, 0.5, use_lower = TRUE)
rsample_sym(A, 0.5, use_lower = FALSE)
}
