% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvd.sam.R
\name{rsvd.sam}
\alias{rsvd.sam}
\title{Compute randomized SVD of a matrix using random sampling}
\usage{
rsvd.sam(A, P, k = max(nu, nv), nu, nv, tol = 1e-05, ...)
}
\arguments{
\item{A}{An input sparse data matrix of type "dgCMatrix". Not necessarily be the adjacency matrix of a network.}

\item{P}{The sampling probability. Should be between 0 and 1.}

\item{k}{Number of singular values requested. \code{k} should not be smaller than \code{nu} or \code{nv}. The default value is
\code{max(nu,nv)}.}

\item{nu}{Number of left singular vectors to be computed.}

\item{nv}{Number of right singular vectors to be computed.}

\item{tol}{Precision parameter of the iterative algorithm. Default is 1e-5.}

\item{...}{Additional arguments of function \code{\link[irlba]{irlba}}.}
}
\value{
\item{u}{The randomized left \code{nu} singular vectors.} \item{v}{The randomized right \code{nv} singular vectors.}
        \item{d}{The \code{k} leading singular values.} \item{sparA}{The sparsified data matrix obtained by \code{rsample(A,P)}.}
}
\description{
Compute the randomized SVD of a matrix by random sampling. The randomized
singular vectors and singular values are computed. Can deal with very large
data matrix.
}
\details{
This function computes the randomized SVD of a data matrix using the random
sampling scheme. The data matrix \code{A} is first sampled to obtain a sparsified matrix.
An iterative algorithm (\code{\link[irlba]{irlba}}) for computing the leading singular vectors is then performed on the
sparsified matrix to obtain the randomized singular vectors and singular values.
}
\examples{

n <- 100
rank <- 2
clustertrue.y <- rep(1:rank, each = n/rank)
clustertrue.z <- rep(1:rank, each = n/rank)
A <- matrix(0, n, n)
for(i in 1:n) {
    for(j in 1:n) {
        A[i, j] <- ifelse(clustertrue.y[i] == clustertrue.z[i], rbinom(1, 1, 0.2), rbinom(1, 1, 0.1))
    }
}
diag(A) <- 0
A <- as(A, "dgCMatrix")
rsvd.sam(A, P = 0.7, nu = rank, nv = rank)

}
\seealso{
\code{\link[RandClust]{rsample}}, \code{\link[irlba]{irlba}}.
}
